+++ examples/lqr.py
@@ -99,9 +99,11 @@ problem.evaluate(xs_i, us_i, prob_data)
 
 solver.setup(problem)
 solver.run(problem, xs_i, us_i)
-res = solver.results
-ws = solver.workspace
+res: aligator.Results = solver.results
+ws: aligator.Workspace = solver.workspace
 print(res)
+print(ws.state_dual_infeas.tolist())
+print(ws.control_dual_infeas.tolist())
 
 plt.subplot(121)
 fig1: plt.Figure = plt.gcf()