+++ include/aligator/modelling/spaces/pinocchio-groups.hpp
@@ -14,39 +14,43 @@ namespace pin = pinocchio;
 /// Type trait. Indicates whether @tparam G is derived from
 /// pinocchio::LieGroupBase.
 template <typename G>
-using is_pinocchio_lie_group = std::is_base_of<pin::LieGroupBase<G>, G>;
+inline constexpr bool is_pinocchio_lie_group =
+    std::is_base_of_v<pin::LieGroupBase<G>, G>;
 
 /// @brief  Wrap a Pinocchio Lie group into a ManifoldAbstractTpl object.
 ///
-template <typename _LieGroup>
-struct PinocchioLieGroup
-    : public ManifoldAbstractTpl<typename _LieGroup::Scalar> {
+template <typename G, std::enable_if_t<is_pinocchio_lie_group<G>> * = nullptr>
+struct PinocchioLieGroup : ManifoldAbstractTpl<typename G::Scalar> {
 public:
-  using LieGroup = _LieGroup;
+  using LieGroup = G;
   using Scalar = typename LieGroup::Scalar;
   using Base = ManifoldAbstractTpl<Scalar>;
+  using Base::ndx;
+  using Base::nx;
   ALIGATOR_DYNAMIC_TYPEDEFS(Scalar);
-  static_assert(is_pinocchio_lie_group<LieGroup>::value,
-                "LieGroup template argument should be a subclass of "
-                "pinocchio::LieGroupBase.");
 
   LieGroup lg_;
-  PinocchioLieGroup() {}
+
+  PinocchioLieGroup()
+      : Base(0, 0) {
+    this->nx_ = lg_.nq();
+    this->ndx_ = lg_.nv();
+  }
   PinocchioLieGroup(const LieGroup &lg)
-      : lg_(lg) {}
-  PinocchioLieGroup(LieGroup &&lg)
-      : lg_(std::move(lg)) {}
+      : Base(lg.nq(), lg.nv())
+      , lg_(lg) {}
   PinocchioLieGroup(const PinocchioLieGroup &lg) = default;
   PinocchioLieGroup(PinocchioLieGroup &&lg) = default;
 
   template <typename... Args>
   PinocchioLieGroup(Args &&...args)
-      : lg_(std::forward<Args>(args)...) {}
+      : Base(0, 0)
+      , lg_(std::forward<Args>(args)...) {
+    this->nx_ = lg_.nq();
+    this->ndx_ = lg_.nv();
+  }
 
-  operator LieGroup() { return lg_; }
-
-  inline int nx() const { return lg_.nq(); }
-  inline int ndx() const { return lg_.nv(); }
+  operator LieGroup() const { return lg_; }
 
   bool isNormalized(const ConstVectorRef &x) const {
     if (x.size() < nx())
@@ -107,6 +111,8 @@ protected:
 
   virtual void rand_impl(VectorRef out) const { out = lg_.random(); }
 };
+
+template <typename G> PinocchioLieGroup(const G &) -> PinocchioLieGroup<G>;
 
 template <int D, typename Scalar>
 using SETpl = PinocchioLieGroup<pin::SpecialEuclideanOperationTpl<D, Scalar>>;