+++ include/aligator/solvers/fddp/solver-fddp.hpp
@@ -36,7 +36,7 @@ template <typename Scalar> struct SolverFDDPTpl {
   using VParams = ValueFunctionTpl<Scalar>;
   using QParams = QFunctionTpl<Scalar>;
   using CostData = CostDataAbstractTpl<Scalar>;
-  using ExpModel = ExplicitDynamicsModelTpl<Scalar>;
+  using DynamicsModel = ExplicitDynamicsModelTpl<Scalar>;
   using ExplicitDynamicsData = ExplicitDynamicsDataTpl<Scalar>;
   using CallbackPtr = shared_ptr<CallbackBaseTpl<Scalar>>;
   using CallbackMap = boost::unordered_map<std::string, CallbackPtr>;
@@ -191,9 +191,6 @@ public:
 
   bool run(const Problem &problem, const std::vector<VectorXs> &xs_init = {},
            const std::vector<VectorXs> &us_init = {});
-
-  static const ExplicitDynamicsData &
-  stage_get_dynamics_data(const StageDataTpl<Scalar> &data);
 };
 
 } // namespace aligator