+++ include/aligator/solvers/proxddp/solver-proxddp.hxx
@@ -17,7 +17,7 @@
 
 namespace aligator {
 
-// [1], realted to Appendix A, details on aug. Lagrangian method
+// [1], related to Appendix A, details on aug. Lagrangian method
 // interpretation as shifted-penalty method
 template <typename Scalar>
 void computeProjectedJacobians(const TrajOptProblemTpl<Scalar> &problem,
@@ -35,8 +35,8 @@ void computeProjectedJacobians(const TrajOptProblemTpl<Scalar> &problem,
     auto &jac = workspace.cstr_proj_jacs[i];
 
     for (size_t j = 0; j < sm.numConstraints(); j++) {
-      jac(j, 0) = sd.constraint_data[j]->Jx_;
-      jac(j, 1) = sd.constraint_data[j]->Ju_;
+      jac(j, 0) = sd.constraint_data[j]->Jx();
+      jac(j, 1) = sd.constraint_data[j]->Ju();
     }
 
     auto Px = jac.blockCol(0);
@@ -54,7 +54,7 @@ void computeProjectedJacobians(const TrajOptProblemTpl<Scalar> &problem,
     auto &jac = workspace.cstr_proj_jacs[N];
     const auto &cds = prob_data.term_cstr_data;
     for (size_t j = 0; j < cds.size(); j++) {
-      jac(j, 0) = cds[j]->Jx_;
+      jac(j, 0) = cds[j]->Jx();
     }
 
     auto Px = jac.blockCol(0);
@@ -97,8 +97,7 @@ void SolverProxDDPTpl<Scalar>::setNumThreads(const size_t num_threads) {
   num_threads_ = num_threads;
   omp::set_default_options(num_threads);
 }
-// [1] Section IV. Proximal Differential Dynamic Programming
-// C. Forward pass
+
 template <typename Scalar>
 Scalar SolverProxDDPTpl<Scalar>::tryLinearStep(const Problem &problem,
                                                const Scalar alpha) {
@@ -142,9 +141,7 @@ Scalar SolverProxDDPTpl<Scalar>::tryLinearStep(const Problem &problem,
   stage.xspace_next_->integrate(results_.xs[nsteps], dx_tmp.head(ndxN),
                                 workspace_.trial_xs[nsteps]);
   TrajOptData &prob_data = workspace_.problem_data;
-  prob_data.cost_ =
-      problem.evaluate(workspace_.trial_xs, workspace_.trial_us, prob_data);
-  return prob_data.cost_;
+  return problem.evaluate(workspace_.trial_xs, workspace_.trial_us, prob_data);
 }
 
 template <typename Scalar>
@@ -215,18 +212,16 @@ bool SolverProxDDPTpl<Scalar>::computeMultipliers(
   const TrajOptData &prob_data = workspace_.problem_data;
   const size_t nsteps = workspace_.nsteps;
 
-  // [1] Section B. Augmented Lagrangian methods eqn. 5a and 5b for x_plus
-  // and in subsection Primal-dual search x_k is prev_x. Then, more precisely
-  // eqn. 39 gives the formula for first-order multiplier estimates in the
-  // DDP setup
+  std::vector<VectorXs> &xs = workspace_.trial_xs;
   std::vector<VectorXs> &lams_plus = workspace_.lams_plus;
 
   const std::vector<VectorXs> &vs_prev = workspace_.prev_vs;
   std::vector<VectorXs> &vs_plus = workspace_.vs_plus;
-  // std::vector<VectorXs> &vs_pdal = workspace_.vs_pdal;
 
   std::vector<VectorXs> &Lvs = workspace_.Lvs;
   std::vector<VectorXs> &shifted_constraints = workspace_.shifted_constraints;
+  std::vector<VectorXs> &stage_infeas = workspace_.stage_infeasibilities;
+  std::vector<VectorXs> &fs = workspace_.dyn_slacks;
 
   assert(Lvs.size() == vs_prev.size());
   assert(Lvs.size() == nsteps + 1);
@@ -234,9 +229,9 @@ bool SolverProxDDPTpl<Scalar>::computeMultipliers(
   // initial constraint
   {
     StageFunctionData &dd = *prob_data.init_data;
-    lams_plus[0] = lams[0] + dd.value_ / mu();
     /// TODO: generalize to the other types of initial constraint (non-equality)
-    workspace_.dyn_slacks[0] = dd.value_;
+    fs[0] = dd.value_;
+    lams_plus[0] = lams[0] + fs[0] / mu();
     RET_FALSE_IF_NAN(lams_plus[0]);
   }
   using ConstraintSetProd = ConstraintSetProductTpl<Scalar>;
@@ -245,15 +240,15 @@ bool SolverProxDDPTpl<Scalar>::computeMultipliers(
   for (size_t i = 0; i < nsteps; i++) {
     const StageModel &stage = *problem.stages_[i];
     const StageData &sd = *prob_data.stage_data[i];
-    const DynamicsData &dd = *sd.dynamics_data;
+    const ExplicitDynamicsDataTpl<Scalar> &dd = *sd.dynamics_data;
     const ConstraintStack &cstr_stack = stage.constraints_;
 
     assert(vs[i].size() == stage.nc());
     assert(lams[i + 1].size() == stage.ndx2());
 
     // 1. compute dynamics error
-    lams_plus[i + 1] = lams[i + 1] + dd.value_ / mu();
-    workspace_.dyn_slacks[i + 1] = dd.value_;
+    stage.xspace_next().difference(xs[i + 1], dd.xnext_, fs[i + 1]);
+    lams_plus[i + 1] = lams[i + 1] + fs[i + 1] / mu();
     RET_FALSE_IF_NAN(lams_plus[i + 1]);
 
     // 2. use product constraint operator
@@ -274,11 +269,14 @@ bool SolverProxDDPTpl<Scalar>::computeMultipliers(
     Lvs[i].noalias() -= mu() * vs[i];
     vs_plus[i] = mu_inv() * vs_plus[i];
     assert(Lvs[i].size() == stage.nc());
+
+    stage_infeas[i] = mu() * (vs_plus[i] - vs_prev[i]);
+    workspace_.stage_cstr_violations[long(i)] =
+        math::infty_norm(stage_infeas[i]);
     RET_FALSE_IF_NAN(Lvs[i]);
   }
 
   if (!problem.term_cstrs_.empty()) {
-    assert(problem.term_cstrs_.size() == prob_data.term_cstr_data.size());
     const ConstraintStack &cstr_stack = problem.term_cstrs_;
     const ConstraintSetProd &op = workspace_.cstr_product_sets[nsteps];
 
@@ -295,8 +293,14 @@ bool SolverProxDDPTpl<Scalar>::computeMultipliers(
     Lvs[nsteps].noalias() -= mu() * vs[nsteps];
     vs_plus[nsteps] = mu_inv() * vs_plus[nsteps];
     assert(Lvs[nsteps].size() == cstr_stack.totalDim());
+
+    stage_infeas[nsteps] = mu() * (vs_plus[nsteps] - vs_prev[nsteps]);
+    workspace_.stage_cstr_violations[long(nsteps)] =
+        math::infty_norm(stage_infeas[nsteps]);
     RET_FALSE_IF_NAN(Lvs[nsteps]);
   }
+  results_.prim_infeas = std::max(math::infty_norm(stage_infeas),
+                                  math::infty_norm(workspace_.dyn_slacks));
   return true;
 }
 
@@ -318,6 +322,7 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::updateGains() {
   VectorRef ff = results_.getFeedforward(N);
   MatrixRef fb = results_.getFeedback(N);
 
+  assert(ff.rows() == linear_solver_->getFeedforward(N).rows());
   ff = linear_solver_->getFeedforward(N).tail(ff.rows());
   fb = linear_solver_->getFeedback(N).bottomRows(fb.rows());
 }
@@ -348,7 +353,7 @@ Scalar SolverProxDDPTpl<Scalar>::tryNonlinearRollout(const Problem &problem,
     // use lams[0] as a tmp var for alpha * dx0
     lams[0] = alpha * dxs[0];
     stage.xspace().integrate(results_.xs[0], lams[0], xs[0]);
-    lams[0] = results_.lams[0] + alpha * workspace_.dlams[0];
+    lams[0] = results_.lams[0] + alpha * dlams[0];
 
     ALIGATOR_RAISE_IF_NAN_NAME(xs[0], fmt::format("xs[{:d}]", 0));
   }
@@ -357,18 +362,15 @@ Scalar SolverProxDDPTpl<Scalar>::tryNonlinearRollout(const Problem &problem,
     const StageModel &stage = *problem.stages_[t];
     StageData &data = *prob_data.stage_data[t];
 
-    const std::array<long, 4> _dims{stage.nu(), stage.nc(), stage.ndx2(),
-                                    stage.ndx2()};
-    BlkMatrix<ConstVectorRef, 4, 1> ff{
-        linear_solver_->getFeedforward(t), _dims, {1}};
-    BlkMatrix<ConstMatrixRef, 4, 1> fb{
-        linear_solver_->getFeedback(t), _dims, {stage.ndx1()}};
-    ConstVectorRef kff = ff[0];
-    ConstVectorRef zff = ff[1];
-    ConstVectorRef lff = ff[2];
+    const std::array<long, 3> dims = {stage.nu(), stage.nc(), stage.ndx2()};
+    BlkMatrix<ConstVectorRef, 3, 1> ff{
+        linear_solver_->getFeedforward(t), dims, {1}};
+    BlkMatrix<ConstMatrixRef, 3, 1> fb{
+        linear_solver_->getFeedback(t), dims, {stage.ndx1()}};
+    ConstVectorRef kff = ff.blockSegment(0);
+    ConstVectorRef zff = ff.blockSegment(1);
     ConstMatrixRef Kfb = fb.blockRow(0);
     ConstMatrixRef Zfb = fb.blockRow(1);
-    ConstMatrixRef Lfb = fb.blockRow(2);
 
     dus[t] = alpha * kff;
     dus[t].noalias() += Kfb * dxs[t];
@@ -378,23 +380,11 @@ Scalar SolverProxDDPTpl<Scalar>::tryNonlinearRollout(const Problem &problem,
     dvs[t].noalias() += Zfb * dxs[t];
     vs[t] = results_.vs[t] + dvs[t];
 
-    dlams[t + 1] = alpha * lff;
-    dlams[t + 1].noalias() += Lfb * dxs[t];
-    lams[t + 1] = results_.lams[t + 1] + dlams[t + 1];
-
-    stage.evaluate(xs[t], us[t], xs[t + 1], data);
-
-    // rollout has zeroed out the residual
-    dyn_slacks[t].setZero();
-
-    if (!stage.hasDynModel() || stage.dynamics_->isExplicit()) {
-      // nothing to do
-    } else {
-      ALIGATOR_RUNTIME_ERROR(
-          "Implicit dynamics are no longer supported (for now)");
-    }
+    stage.evaluate(xs[t], us[t], data);
+    xs[t + 1] = data.dynamics_data->xnext_;
 
     stage.xspace_next().difference(results_.xs[t + 1], xs[t + 1], dxs[t + 1]);
+    lams[t + 1] = results_.lams[t + 1] + alpha * dlams[t + 1];
 
     ALIGATOR_RAISE_IF_NAN_NAME(xs[t + 1], fmt::format("xs[{:d}]", t + 1));
     ALIGATOR_RAISE_IF_NAN_NAME(us[t], fmt::format("us[{:d}]", t));
@@ -422,7 +412,7 @@ Scalar SolverProxDDPTpl<Scalar>::tryNonlinearRollout(const Problem &problem,
         linear_solver_->getFeedforward(nsteps), _dims, {1}};
     BlkMatrix<ConstMatrixRef, 2, 1> fb{
         linear_solver_->getFeedback(nsteps), _dims, {ndx}};
-    ConstVectorRef zff = ff[1];
+    ConstVectorRef zff = ff.blockSegment(1);
     ConstMatrixRef Zfb = fb.blockRow(1);
 
     dvs[nsteps] = alpha * zff;
@@ -584,7 +574,6 @@ bool SolverProxDDPTpl<Scalar>::innerLoop(const Problem &problem) {
     ALIGATOR_TRACY_ZONE_NAMED_N(FilterPairEval, "pair_eval_fun", true);
     std::pair<Scalar, Scalar> fpair;
     fpair.first = forwardPass(problem, a0);
-    computeInfeasibilities(problem);
     fpair.second = results_.prim_infeas;
     return fpair;
   };
@@ -614,7 +603,6 @@ bool SolverProxDDPTpl<Scalar>::innerLoop(const Problem &problem) {
     if (force_initial_condition_) {
       workspace_.Lxs[0].setZero();
     }
-    computeInfeasibilities(problem);
     computeCriterion();
 
     // exit if either the subproblem or overall problem converged
@@ -626,14 +614,8 @@ bool SolverProxDDPTpl<Scalar>::innerLoop(const Problem &problem) {
     computeProjectedJacobians(problem, mu_inv(), workspace_);
     this->initializeRegularization();
     this->updateLQSubproblem();
-    // TODO: supply a penalty weight matrix for constraints
 
-    // In the next two lines, the LQ subproblem is solved. This is
-    // another way to view the backward and forward passes of the
-    // Riccati recursion detailed in [1] (IV. Proximal Differential
-    //  Dynamic Programming Section B Backward). The backward pass
-    // computes the gains, and the forward pass computes the new
-    // control and state trajectories.
+    // Solve the LQ subproblem.
     linear_solver_->backward(mu());
 
     linear_solver_->forward(workspace_.dxs, workspace_.dus, workspace_.dvs,
@@ -659,7 +641,6 @@ bool SolverProxDDPTpl<Scalar>::innerLoop(const Problem &problem) {
     switch (sa_strategy_) {
     case StepAcceptanceStrategy::LINESEARCH_ARMIJO:
     case StepAcceptanceStrategy::LINESEARCH_NONMONOTONE:
-      assert(linesearch_.isValid());
       phi_new = linesearch_.run(merit_eval_fun, phi0, dphi0, alpha_opt);
       break;
     case StepAcceptanceStrategy::FILTER:
@@ -709,36 +690,6 @@ bool SolverProxDDPTpl<Scalar>::innerLoop(const Problem &problem) {
   return false;
 }
 
-template <typename Scalar>
-void SolverProxDDPTpl<Scalar>::computeInfeasibilities(const Problem &problem) {
-  ALIGATOR_NOMALLOC_SCOPED;
-  ALIGATOR_TRACY_ZONE_SCOPED;
-  const size_t nsteps = workspace_.nsteps;
-
-  std::vector<VectorXs> &vs_plus = workspace_.vs_plus;
-  std::vector<VectorXs> &vs_prev = workspace_.prev_vs;
-  std::vector<VectorXs> &stage_infeas = workspace_.stage_infeasibilities;
-
-  // compute infeasibility of all stage constraints [1] eqn. 53
-  for (size_t i = 0; i < nsteps; i++) {
-    stage_infeas[i] = mu() * (vs_plus[i] - vs_prev[i]);
-
-    workspace_.stage_cstr_violations[long(i)] =
-        math::infty_norm(stage_infeas[i]);
-  }
-
-  // compute infeasibility of terminal constraints
-  if (!problem.term_cstrs_.empty()) {
-    stage_infeas[nsteps] = mu() * (vs_plus[nsteps] - vs_prev[nsteps]);
-
-    workspace_.stage_cstr_violations[long(nsteps)] =
-        math::infty_norm(stage_infeas[nsteps]);
-  }
-
-  results_.prim_infeas = std::max(math::infty_norm(stage_infeas),
-                                  math::infty_norm(workspace_.dyn_slacks));
-}
-
 template <typename Scalar> void SolverProxDDPTpl<Scalar>::computeCriterion() {
   ALIGATOR_NOMALLOC_SCOPED;
   ALIGATOR_TRACY_ZONE_SCOPED;
@@ -754,12 +705,16 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::computeCriterion() {
     Scalar rd = math::infty_norm(workspace_.dyn_slacks[i]);
     // path constraints
     Scalar rc = math::infty_norm(workspace_.Lvs[i]);
+    // if (rollout_type_ == RolloutType::NONLINEAR)
+    //   rx = 0.0;
 
     workspace_.stage_inner_crits[long(i)] = std::max({rx, ru, rd, rc});
     workspace_.state_dual_infeas[long(i)] = rx; // [1] eqn. 52
     workspace_.control_dual_infeas[long(i)] = ru;
   }
   Scalar rx = math::infty_norm(workspace_.Lxs[nsteps]);
+  // if (rollout_type_ == RolloutType::NONLINEAR)
+  //   rx = 0.0;
   Scalar rc = math::infty_norm(workspace_.Lvs[nsteps]);
   workspace_.state_dual_infeas[long(nsteps)] = rx; // [1] eqn. 52
   workspace_.stage_inner_crits[long(nsteps)] = std::max(rx, rc);
@@ -770,17 +725,12 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::computeCriterion() {
                math::infty_norm(workspace_.control_dual_infeas));
 }
 
-template <typename Scalar>
-void SolverProxDDPTpl<Scalar>::registerCallback(const std::string &name,
-                                                CallbackPtr cb) {
-  callbacks_.insert_or_assign(name, cb);
-}
-
 template <typename Scalar> void SolverProxDDPTpl<Scalar>::updateLQSubproblem() {
   ALIGATOR_NOMALLOC_SCOPED;
   ALIGATOR_TRACY_ZONE_SCOPED;
   gar::LqrProblemTpl<Scalar> &prob = workspace_.lqr_problem;
   const TrajOptData &pd = workspace_.problem_data;
+  const auto &dyn_slacks = workspace_.dyn_slacks;
 
   using gar::LqrKnotTpl;
 
@@ -792,15 +742,15 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::updateLQSubproblem() {
     // taking the knot and not a view, since resizes might
     // happen here.
     LqrKnotTpl<Scalar> &knot = prob.stages[t];
-    const DynamicsData &dd = *sd.dynamics_data;
+    const ExplicitDynamicsDataTpl<Scalar> &dd = *sd.dynamics_data;
     const CostData &cd = *sd.cost_data;
     uint nx = knot.nx;
     uint nu = knot.nu;
     uint nc = knot.nc;
 
-    knot.A = dd.Jx_;
-    knot.B = dd.Ju_;
-    knot.f = dd.value_;
+    knot.A = dd.Jx();
+    knot.B = dd.Ju();
+    knot.f = dyn_slacks[t + 1];
 
     knot.Q = cd.Lxx_;
     knot.S = cd.Lxu_;
@@ -818,7 +768,6 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::updateLQSubproblem() {
       knot.R += dd.Huu_;
     }
 
-    // TODO: handle the bloody constraints
     assert(knot.nc == workspace_.cstr_proj_jacs[t].rows());
     knot.C.topRows(nc) = workspace_.cstr_proj_jacs[t].blockCol(0);
     knot.D.topRows(nc) = workspace_.cstr_proj_jacs[t].blockCol(1);
@@ -842,7 +791,7 @@ template <typename Scalar> void SolverProxDDPTpl<Scalar>::updateLQSubproblem() {
   }
 
   const StageFunctionData &id = *pd.init_data;
-  prob.G0 = id.Jx_;
+  prob.G0 = id.Jx();
   prob.g0 = id.value_;
 
   LqrKnotTpl<Scalar> &model = prob.stages[0];