+++ include/aligator/utils/forward-dyn.hpp
@@ -1,5 +1,5 @@
 /// @file forward-dyn.hpp
-/// @copyright Copyright (C) 2022-2024 LAAS-CNRS, INRIA
+/// @copyright Copyright (C) 2022-2024 LAAS-CNRS, 2022-2025 INRIA
 #pragma once
 
 #include "aligator/core/explicit-dynamics.hpp"
@@ -8,14 +8,13 @@
 
 namespace aligator {
 
-/**
- * @brief    Evaluates the forward map for a discrete dynamics model, implicit
- * or explicit.
- * @details  If the given DynamicsModelTpl can be safely downcast to an explicit
- * dynamics type then this function will use the
- * ExplicitDynamicsModelTpl::forward() method.
- */
-template <typename T> struct forwardDynamics {
+template <typename T> struct ForwardDynamicsOptions {
+  uint max_iters = 1000u;
+  T eps = 1e-6;
+};
+
+namespace detail {
+template <typename T> struct ForwardDynamics {
 
   using VectorXs = typename math_types<T>::VectorXs;
   using VectorRef = typename math_types<T>::VectorRef;
@@ -24,46 +23,53 @@ template <typename T> struct forwardDynamics {
 
   static void run(const DynamicsModelTpl<T> &model, const ConstVectorRef &x,
                   const ConstVectorRef &u, DynamicsDataTpl<T> &data,
-                  VectorRef xout,
-                  const std::optional<ConstVectorRef> &gap = std::nullopt,
-                  const uint max_iters = 1000, const T EPS = 1e-6) {
-    using ExpModel = ExplicitDynamicsModelTpl<T>;
-    using ExpData = ExplicitDynamicsDataTpl<T>;
-
-    if (model.isExplicit()) {
-      const ExpModel &model_cast = static_cast<const ExpModel &>(model);
-      ExpData &data_cast = static_cast<ExpData &>(data);
-      run(model_cast, x, u, data_cast, xout, gap);
-    } else {
-      // create NewtonRaph algo's data
-      VectorXs dx0buf(model.ndx2);
-      dx0buf.setZero();
-      NewtonRaphson<T>::run(
-          model.space_next(),
-          [&](const ConstVectorRef &xnext, VectorRef out) {
-            model.evaluate(x, u, xnext, data);
-            out = data.value_;
-            if (gap.has_value())
-              out += *gap;
-          },
-          [&](const ConstVectorRef &xnext, MatrixRef Jout) {
-            model.computeJacobians(x, u, xnext, data);
-            Jout = data.Jy_;
-          },
-          x, xout, data.value_, dx0buf, data.Jy_, EPS, max_iters);
-    }
+                  VectorRef xout, const std::optional<ConstVectorRef> &gap,
+                  ForwardDynamicsOptions<T> opts) {
+    // create NewtonRaph algo's data
+    VectorXs dx0buf(model.ndx2);
+    dx0buf.setZero();
+    NewtonRaphson<T>::run(
+        model.space_next(),
+        [&](const ConstVectorRef &xnext, VectorRef out) {
+          model.evaluate(x, u, xnext, data);
+          out = data.value_;
+          if (gap.has_value())
+            out += *gap;
+        },
+        [&](const ConstVectorRef &xnext, MatrixRef Jout) {
+          model.computeJacobians(x, u, xnext, data);
+          Jout = data.Jy_;
+        },
+        x, xout, data.value_, dx0buf, data.Jy_, opts.eps, opts.max_iters);
   }
 
   static void run(const ExplicitDynamicsModelTpl<T> &model,
                   const ConstVectorRef &x, const ConstVectorRef &u,
                   ExplicitDynamicsDataTpl<T> &data, VectorRef xout,
-                  const std::optional<ConstVectorRef> &gap = std::nullopt) {
+                  const std::optional<ConstVectorRef> &gap,
+                  ForwardDynamicsOptions<T>) {
     model.forward(x, u, data);
     xout = data.xnext_;
-    if (gap.has_value()) {
+    if (gap.has_value())
       model.space_next().integrate(xout, *gap, xout);
-    }
   }
 };
+} // namespace detail
+
+/// @brief    Evaluates the forward map for a discrete dynamics model, implicit
+/// or explicit.
+/// @details  If the given model is explicit then this function will just use
+/// the ExplicitDynamicsModelTpl::forward() method.
+template <typename T, template <typename> class M,
+          typename D = typename M<T>::Data>
+void forwardDynamics(const M<T> &model,
+                     const typename math_types<T>::ConstVectorRef &x,
+                     const typename math_types<T>::ConstVectorRef &u, D &data,
+                     const typename math_types<T>::VectorRef xout,
+                     const std::optional<typename math_types<T>::ConstVectorRef>
+                         &gap = std::nullopt,
+                     ForwardDynamicsOptions<T> opts = {}) {
+  detail::ForwardDynamics<T>::run(model, x, u, data, xout, gap, opts);
+}
 
 } // namespace aligator