+++ tests/problem.cpp
@@ -38,8 +38,8 @@ struct MyModel : ExplicitDynamicsModelTpl<double> {
 
   void dForward(const ConstVectorRef &x, const ConstVectorRef &u,
                 ExplicitData &data) const {
-    space_next().Jintegrate(x, u, data.Jx_, 0);
-    space_next().Jintegrate(x, u, data.Ju_, 1);
+    space_next().Jintegrate(x, u, data.Jx(), 0);
+    space_next().Jintegrate(x, u, data.Ju(), 1);
   }
 };
 
@@ -102,7 +102,6 @@ TEST_CASE("test_problem", "[node]") {
   auto nu = f.nu;
   auto &space = f.space;
   const auto &stage = *f.problem.stages_[0];
-  REQUIRE(stage.numPrimal() == space.ndx() + nu);
   REQUIRE(stage.numDual() == space.ndx());
 
   auto *p_dyn = stage.getDynamics<MyModel>();
@@ -123,8 +122,8 @@ TEST_CASE("test_problem", "[node]") {
 
   fmt::print("{}\n", stage);
 
-  auto stage_data = stage.createData();
-  stage.evaluate(x0, u0, x0, *stage_data);
+  shared_ptr stage_data = stage.createData();
+  stage.evaluate(x0, u0, *stage_data);
   REQUIRE(stage_data->cost_data->value_ == 0.);
 
   TrajOptDataTpl<double> prob_data(f.problem);